Add the submodule "jython"
This adds a recipe to build Jython.

But beware!  You cannot run commit-submodule.py on this submodule, as
the script is written in Jython, and during the execution, jython.jar is
deleted.  Catch-22.

Therefore, it is implemented as a shell script (since we need to inject
Lib/ into the .jar).

Oh, and you need a Python 2.5 in /usr/lib/python2.5/ before compiling
(which is the default with Ubuntu installations).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>