Fiji build: when copying a .jar, check time, not hashCode
For the up-to-date check (if a source .jar is newer than a target .jar)
we used to rely on JarEntry's hashCode() method.

This method proved to be unreliable.

As we do not track the destination files in Git anyway, it is better to
check the time, and copy _even_ if the .jar files contain identical
files (apart from the timestamps).

Noticed by Stephan Saalfeld.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>