Fiji build: fix prefixing of SubFakes' CLASSPATH variables
In submodules, you can set the CLASSPATH in the Fakefiles, but as Java
has no concept how to change directories (and as the CLASSPATH could
already contain .jar files to be found in the parent directory), we have
to prefix them with the correct (relative) path.

This used to work until I had to revert the prefixing with an absolute
path, as that broke Windows support.

This commit introduces relative path prefixes, and everything should be
kosher again in the state of Denmark.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>