Adjusted LevelSet.java for usage without a GUI.
 * Replaced spaces by underscores in GenericDialog entries to deliver
   unique names for macros that fill in values.
 * Added method setParameters to set ALL parameters that would otherwise
   be set from the GenericDialog.
 * The boolean ask_params is used as well to prevent painting a progress
   image when not set.

Still unresolved: some variables of the GenericDialog have state and
thus are preserved across multiple instantiations of the plugin, other
variables do not. The ones with state belong to the FastMarching class,
and are stored there as static fields. ALL should be preserved or none,
the current situation is very confusing.