Add a script to (cross) compile tijmp, a memory profiler for Java6
Memory profiling on Java6 has become substantially easier with the
advent of 'tijmp'.  To find out what is hogging your memory, click
on the 'Walk heap' button.  For every listed data type, you will
now be able to show the owners.

Call it with:

	$ ./fiji -Dtijmp.jar=tijmp/tijmp.jar -agentlib:tijmp/tijmp --

Hopefully, we can contribute to the usefulness of 'tijmp' by
extending its capabilities, e.g. adding a timestamp to the owners,
or a lifetime filter.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>