Introduce main Fiji class
Instead of launching ImageJ directly, we will launch fiji.Main, which
registers a listener to set icon and title of all frames, before
calling ImageJ's main method.

This solves two issues:

- we no longer rely on a patched ImageJA to be able to set the icon and
  window title of ImageJ's main window, and

- we set the icon of all windows, even Swing windows (e.g. the 3D viewer).

It should also solve the odd timing-related issue when loading an image
and loading the icon blocked each other.

Further, the main class opens the door to play games with ASM, rewriting
parts of ImageJ on the fly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>