tests/lib.py: make use of fiji.Main.waitForWindow()
The waitForFrame function was racy, and besides, it could not wait for
a dialog.  So use the newly introduced waitForWindow() method in
fiji.Main.  This needs a one-line change in tests/3d.py.

While at it, add a few more helper functions.

So now you can wait for a dialog, press a button in the latest
waited-for dialog, and click a menu item in the frame that was last
waited for.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>