Add a simple interface for reporting bugs in Fiji.
It has been suggested that the Bugzilla "new bug" page
makes people disinclined to report bugs, so hopefully
this provides a simpler way of creating these reports.

The plugin gathers some useful information
from the Java System properties and prepends the bug
description with them.

The "Report a Bug" option appears near the bottom of
the "Help" menu.

It would be easy to bring up this plugin complete
with a stack trace on catching particular Throwables
(e.g. NPEs) from plugins, but this may not be desirable.