Script to generate wiki page from plugins folder
'python scripts/plugin-list-parser.py' parse the plugins folder for
scripts, macros and plugins, and generate a pretty output, using
Wikimedia tags for direct output to the Fiji wiki (for instance).
The output is sorted by plugin location within the Fiji menu system.

If it finds a jar, it will try to look for a .config file in the
staged-plugins folder to infer the plygins location in the Fiji menu
system. You can check a result of the ouput on the Test.txt page of the
Fiji wiki.

Signed-off-by: tinevez <tinevez@intel-tinevez-3-302.mpi-cbg.de>

*	new file:   scripts/plugin-list-parser.py