New script to interrogate documentation from the wiki
scripts/plugin-documentation-list.py contains facilities to remotely
interrogate the wiki about documented or un-documented plugins.

The current version create a blame page when run, that reports all the
plugins (in the Plugin Menu web page, generated by another script) that
have no documentation or no maintainer.
It was use to create a Blame Page that can be seen here:

http://pacific.mpi-cbg.de/wiki/index.php/Plugins_documentation_todo

Signed-off-by: Jean-Yves Tinevez <tinevez@mpi-cbg.de>