Made antialiased text selection function more reliable.
There was a unreliability problem with the selection of anti aliased text.
Sometimes it just didnt seem to listen to the GUI setting.
It is set with a GUI check box returning a boolean.
This was being used in the run method to do ip.setAntialiasedText(AAtext);
but i moved that and ip.setColor(Toolbar.getForegroundColor());
to the showDialog where other font related stuff is set in  ip.methods
Seems to behave more reliably now (but maybe still not alaways as expected)