AbstractInterpreter: work around JTextArea.getColumns() being buggy
JTextArea's getColumns() method is off when the text area lives inside
a JScrollPane which does not show its vertical scroll bar.  This is
annoying, as there is no other way to determine the current column than
to use the information from getColumns() and from getLineStartOffset().

So just force the scroll bar to be visible at all times.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>