Script to get who maintains what.
The script scripts/plugin-documentation-list.py can now generate a wiki
page that will collect the list of maintainers and their plugins and
pakckages.

You can see the result on the wiki:
http://pacific.mpi-cbg.de/wiki/index.php/Maintainer_page

The data is fetch from the Wiki. By default, the script generate wiki markup
that should be put here:

http://pacific.mpi-cbg.de/wiki/index.php/Template:Maintainers

Signed-off-by: Jean-Yves Tinevez <tinevez@mpi-cbg.de>