Fiji launcher: when retrying with less memory, prepend the option
When the main class is not ImageJ, but another class, we stop parsing
the command line parameters after the '--' and pass the rest verbatim to
the main() method.

That did not work well with our method of appending a --mem option
specifying less memory.

So prepend it, and strip out the --mem options.

Noticed by Yap Chin Kiet.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>