moved timeString into its own method, changed == to .equals for string comparisons, typos and commments
the code to make the timeString was part of the run method,
so i took it out and make its own method out of it.
(intention is to impreoive readability, developmentability and OO of code)
Changes string comparisons for digitaiOrDecimal from == to .equals,
as that is more rubust against futiure changes, where strings might
be constructed on the fly.
Added more comments and fixed typos in comments.
AAtext still doesnt work properly:
if i dont change the font size by typing in the GUI - it ignores the AA text setting!!! Why???