Add the plugin "Tutorial Maker"
Given a page title, a wiki login and password, this plugin should make
it much easier for everybody to write tutorial pages.

The idea is to write a description while performing the actions you
describe, from time to time taking a snapshot (which you can crop,
annotate, modify to your heart's wishes).

To allow screenshots of opened menus, you can trigger a snapshot by
hovering for 1 second on the small, or 3 seconds on the large button.
Or click on the large button, which leaves you three seconds to open
the menu.

This version does not upload yet, neither does it have a preview option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>