made maxwidth a separate method, instead of on the fly cpomputed in the run method
took maxwith calculation out of the run method and made it
live in its own method.
takes arguments ImageProcessor ip, double startTime, double intervalTime, int lastFrame
Maxwidth is the maxwidth of the time stamp string,
so we can make sure the time stamp doesnt fall off the right side of the image.
we use it to adjust x and y where the time stamp is written.