Ok, so its very broken, but added preview function using EnhancedPluginFilter
now using EnhancedPluginFilter to make use of preview checkbox
and gd DialogListener functionality.
Moved gd field listening code to required new method
public boolean dialogItemChanged(GenericDialog gd, AWTEvent e)
So i dint need to make my own event listerner gadgets,
(like the scale bar plugin does)
as it's abstracted here already for me to use.

To showDialog method, added gd.PreviewCheckBox  and gd.DilaogListener
and made it retuen the plugin flags eg STACK_REQUIRED etc. as it should.
Now showDialog method takes three arguments: imp (nor ip!), command and
a PluginFilterRunner object called pfr. I guess that comes from the IJ
system that runs the plugin from the menu selection etc.
Was not sure how to get that object, but saw in other
enhancedpluginfilter plugin that it can be declared as a variable
alsong with the others, with default value, which i guess defults
to the real pfr that runs the plugin.
pfr is the argument for the gd.PreviewCheckBox .
Seems to work ok like that, but i dont really understand why...

Using EnhancedPluginFilter, showDialog is run after setup method, but before the run method.

Also we need a
public void setNPasses(int nPasses)
method that works out how many times the run method
will be executed by the plugin when one hits OK.
Dont know why thats needed....? But it is required
for enhancedpluginfilter. Not sure what implications it has
for the line:
if (canceled || frame<first || frame>last) return;
Hopefully none, so long as i calculate nPasses correctly!

Preview chock box = true executes the run method,
so it increments frame variable if you click it off and on again,
which is not what we want. Thr time stamps are then wrong,
as frame is already set too high. Rather we want to use the longest
TimeStamp string to be previewed (which we know from maxWidth method)
and for it not to touch the frame variable. That means i need to
override  the preview checkbox functionality that executes run,
and make it execute a new method that just draws the longest timestamp
into the image, and leaves frame variable alone.

I put the font size and location code into their own
methods each, to get that code into objects outside the run method,
which are called by the run method.
This is so i can begin to debug why the antialiasing functionality
does not work as expected. But, this reults in the font size only being
correct for the first frame (ROI size works still),
but size = 80 for the rest...
which is the size it set as a maximum... Not sure how it gets set to
that... for the second and subsequent frames the vaule of size
must be defaulted to something odd and large to get reset to 80?