Add an initial version of a LiveCD creator script
This script is able to generate a live CD on any recent Ubuntu/Intel
system.  It does so by creating a chroot'ed environment, and using the
live-helper package.

This works only with i386 so far, need to investigate how to
support 64-bit (x86_64, _not_ ia64).  Probably just a question of
providing '-a x86_64', but that will only work on 64-bit systems.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>