Get rid of the last unbundled plugins
To control into which menus plugins go, they have to live in a .jar file
with an appropriate plugins.config file.

So let's get rid of the last .class plugins; Analyze>Grid is a programming
example and should not live in the distributed version anyway, and
HandleExtraFileTypes has its natural place in IO_.jar (and is not a plugin
in the first place).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>