common.RefreshScripts: fix stripping out plugin .jar files
To make running scripts consistent with the PluginClassLoader, which
loads the plugin .jar files only when they could not be found by the
system class loader, we need to add the plugin .jar files to the end
of the class path in RefreshScripts.

Unfortunately, the stripping of the plugin .jar files before they
were re-appended was broken.  This patch fixes that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>