fixed setNPasses method for preview and added check for font size ROI interaction
In SetNPasses method, setting
frame = last;
makes the preview gadget not preview a timestamp,
but it workas as expected wen set to first.

time is set to lasttime() so the longest timestampt to be made is previewed.

in setFontParams method added an if test to see of the ROI
is as big as the image, which is what happens when no ROI is manually set.
There is always an ROI in an ImageProcessor object.
If the ROI is not the same size as the image, then we will
continue to use the font size set in the plugin GUI.
This will fail of one does select all, as in that case
the user has made a ROI, btu its size is ignored and the GUI entered sizie is used.