Fiji Updater: Test if ij.jar is writable
In some installations, the user cannot overwrite ij.jar, and therefore
it is unlikely that _any_ file is writable by the user.  Give a fat
warning in that case, and a hint that you might need to be administrator
for the update to succeed, and ask if we should continue.

Noticed by Arnim Jenett and Albert Cardona.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>