Fiji launcher: when updating, 0-byte files mean "remove me"
Upon startup, the Fiji launcher checks if there is an "update/" directory
in the Fiji root directory, and if there is, moves the files over,
eventually removing the "update/" directory itself.

This procedure is necessary as you cannot replace, say, ij.jar in-flight:
the operation would succeed, but the Java Runtime would get utterly
confused, so much so that at times, I could not even quit ImageJ
properly.

This mechanism did not have anything in place for _removing_ files,
though.

So just use 0-byte files as placeholders for files that need removing.

This also helps the transition during the time when the user does not
update the Fiji launcher itself (which is necessary to actually remove
the files): ImageJ's PluginClassLoader (even the new, URLClassLoader
based one) and the Java Runtime just ignore 0-byte .jar files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>