Add Fabrice Cordelières' v2.0 3d object counter plugin
This is the most modern and functional version of the 3d object counter
I could find.  Dating from 2007, it is from the mac biophotonics
collection, and is newer than the 1.5 version from J. Jackson.  The
plugin is put in the menu under analysis, as per the added .config
file.

There is a small bug: if you type a number less than 50 into the
threshold field of the gui, it makes it fifty something... you have to
use the arrows to click the number down to what you want.

[UTF-8'ified and unDOS-ified by Johannes]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>