Refresh Javas: do not force-load javac's main class
Earlier, javac was loaded using the PluginClassLoader's option to
force (re)loading the class.  This was misguided, as javac is
already in the classpath and does not change all that often.

So just load the class normally and be done with it.

This also happens to fix the compile problem with newer ImageJs,
where the class loader is now based on the URLClassLoader, and you
cannot force reloading a class anymore.

Noticed by Stephan Saalfeld.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>