Fix empty argument for MacOS X Fiji.app
When Fiji.app is double clicked it runs with startup dir as root ("/")
and with an extra empty argument that ImageJ tries to open as the
root directory since File("") gives the root dir.

This mac-specific patch checks for the presence of the extra empty
argument, which is diagnostic of being double clicked rather than being
run directly from the command line and then throws away the empty
argument.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>