StartupMacros.txt: use getDirectory("luts")
This is a fall-out of the recent discussion about
getDirectory("startup"): many macros use that command to determine
the root directory of ImageJ.  This is wrong, as per the documentation:

	getDirectory("startup") - Returns the path to the directory that
	ImageJ was launched from (usually the ImageJ directory).

So it _can_ return something different from the ImageJ root directory.

The correct way, of course, is to use getDirectory("luts"), which is
available since ImageJ v1.37c, 4 April 2006.

So let's set an example and use the correct way ourselves.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>