Tutorial Maker: add Preview option
Now you can click Wiki>Preview and a preview of the page as it will
(hopefully) look like on the Wiki will be opened.

The idea is to use the Wiki's own preview option and parse what it
outputs to show only the preview, with the images redirected to
temporary files (if the images have not been saved yet).

The only drawback I can see so far is that the images, which are not
yet uploaded, cannot be shown resized.  That is something the
user will have to touch up on the Wiki itself.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>