Tutorial Maker: add a toggle to put snapshots into the background
When showing menu items (which is quite frequent a necessity in ImageJ),
it is better to leave the snapshots in the background and to continue
doing the stuff you want to demonstrate.  So make that the default.

In this scenario, it is best to write the text and to crop the snapshots
appropriately in the end.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>