Add the Script Editor
This adds the script editor, as coded during the Google Summer of Code
project of the same name in 2009.

The script editor's purpose is to make it easy to write scripts for Fiji in
Jython, JRuby, Clojure, Javascript, BeanShell and Java.  Syntax
highlighting and completion of class names are the strong points of the
editor.

At the moment, a rudimentary Java debugger is implemented already, but it
needs the JDK's tools.jar to be in the classpath.  This will be fixed in the
future, and likewise, support will be added for ImageJ's macro language,
and we will investigate how feasible support for debugging scripts is.

The commit was made from 5f1daf187fe7182f663499b46cd73f2bed2636ef, which
is available in fiji-historical.git, using these commands:

	git diff --binary -M ...5f1daf187fe7182f663499b46cd73f2bed2636ef |
	git apply --index --whitespace=fix

and then removing some unneeded files, as well as cleaning up patches
outside of src-plugins/Script_Editor/.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>