Add scripts/update-fiji.py to update packages directly on pacific
Earlier, we required developers to call

	fiji --main-class=UpdateFiji update <file> [<files>...]

to put new versions into the updater.  But the class was moved, and
in the distant future, we will use fiji.pluginManager.utilities.PluginData
(phew, what a long name) directly.

So let's have a script now that will not move and continue to work whatever
we change elsewhere:

	fiji scripts/update-fiji.py <file> [<files>...]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>