Move the last sources into src-plugins/<jar-basename>/
There is a convention which will come in quite handy when we will finally
support Netbeans and Eclipse: all .jar sources live in a subdirectory
of src-plugins/ whose name is identical to the basename of the .jar file.

Example: the sources for jars/fiji-scripting.jar are in
src-plugins/fiji-scripting/.

Let's make things consistent and fix the last violator of this rule.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>