Plugin Manager's Updater: prepare for command line usage
A lot of places in the Plugin Manager's source code seem to be limited
to GUI use, even if the package structure seems to suggest otherwise.

Prepare at least the Updater for non-GUI use, so that we can push updates
via the command line (which is needed for full releases).

This patch also shows beautifully one of the biggest curses of software
development: side effects.

One side effect of instantiating the Updater using a Plugin Manager
instance is that it resets the status of _all plugins_!  I.e. if a plugin
was marked "to upload" before, it is now reset to "no change".  Worse,
this behavior was even intended, as the only way the updater will know
about plugins is when they were marked "to upload" to begin with!

It took this developer three hours to hunt down the bug, which was only
visible when instantiating the Updater from a command line (which is the
whole purpose of the current patch series).

This is a serious flaw, and will need proper auditing at a later stage.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>