Plugin Manager: refactor FileUploader
Unfortunately, the FileUploader was designed with several purposes in mind;
that makes it impossible to use it for just _one_ purpose.

So let's at least refactor out the SSH stuff, which will not be needed
(or better said, which is not wanted at all) for the command line updater;
we want to be able to make full releases (or just tiny updates) on the
server itself, via command line.

The base class brings back the functionality to "upload" locally (we had
that already during the development of the Plugin Manager, but this
functionality was thrown away for no good reason).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>