Plugin Manager: clean up utilities.Downloader
Rather than have a complicated jungle of classes that somehow all contribute
to downloading files and showing the progress, let's have some nice, concise
and elegant framework.

That means that we refrain from having several home-baked implementations
of Observer/Observable when Java has a perfectly fine implementation
already.

That also means that we do not try to put into functions into a class that
do not logically belong there but somewhere else.

And it means that we do not hold redundant information all over the place,
but exactly once.  That way we do not risk inconsistent data, and, oh, we
do not risk having many bugs to begin with, as they do not have many rocks
to hide under.

All this results not only in cleaner code, but also in a nice code
reduction.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>