Plugin Manager: rework plugin filtering
Now we no longer carry copies of filtered plugin collections around
anymore, but filter one (and only one) instance all the time.

By this method, stacked filters are possible, such as searching only
in the installed plugins.

While at it, change the unintuitive "search in all metadata" to
a "search only in the filename" paradigm.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>