Plugin Manager: add action buttons
It is not an exactly intuitive user interface to present a list of plugins
with their current states, expecting the user to guess that a click on the
state will open a combo box from which to choose the desired action.

So add buttons "Keep as-is", "Update" / "Install" and "Remove" buttons.

While at it, do some more cleanups on the go.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>