Use a better she-bang line in bin/ and tests/
Rather than having #!/usr/bin/python in the first line, use

	#!/bin/sh
	''''exec "$(dirname "$0")"/../fiji --jython "$0" "$@" # '''

Explanation: Jython interprets everything enclosed in pairs of ''' as
multi-line strings.  This is the common way to comment-out code in
Python.

For /bin/sh, everything between single quotes is considered
verbatim text (i.e. no variables are expanded), so shell interprets the
command on the second line as

	exec <fiji-root>/fiji --jython <original-command-line>

Finally, shell ignores everything after a #, so that we can close the
string for Jython without the shell getting confused.

With this, you finally can say

	./bin/update-fiji.py plugins/TrakEM2.jar

without Python complaining that it did not find the java classes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>