Fiji Updater: use synchronize-db.xml.gz.py for a different purpose
Now that db.xml.gz has been synchronized, and the update-fiji.py that
uses the new Fiji Updater hit 'master', we can use the same script
for a different purpose: update db.xml.gz to contain the checksums
of the files of previous releases.

You have to call the script with a directory that contains the
previous release for now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>