nightly-build.sh: make more idiot proof
Recently, this developer started to use the nightly-build.sh script as
a convenient method to compile the current master in a pretended fresh
checkout.  This mode was triggered by passing a parameter, which is the
name of the commit to test.

But it is possible to forget to pass that parameter.

As the result -- which happened to the particular idiot who wrote this
commit message -- is a cleaned-out working directory, the consequences
are quite dramatic.

So let's make extra-sure that the working directory in which the script
was run is named "nightly-build", and default to testing "HEAD"
otherwise.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>