nightly-build.sh: use Java of parent if no Java submodule was checked out
The nightly-build.sh script can be used to quickly test-compile pretending
to have a fresh checkout.  But we need a JDK to compile, since we have
the script editor with beginnings of debug support now.

So let's just reuse the parent directory's Java if none was found inside
the nightly build's java/ subdirectory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>