Fiji Build: loosen the up-to-date check a little
We used to say that a target is not up-to-date when it had the same
mtime as one source file.  The rationale was that the source could
have been modified in the same split second as the target, but still
later.

However, this is highly unlikely, as we are compiling Java.  Plus,
we'd need to re-run the Fiji Build within the same split second, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>