Fiji Build: refactor to have one common getRule(String prereq)
The prerequisites can contain something like "bla.jar/", which means
that we want to include the contents rather than the whole .jar as
a single file.  Still, when there is a rule for "bla.jar", we want
to build that first.

In the future, there might be other instances where the prerequisite
string differs from the rule string.

So let's just have a getRule(String prereq) method, which would be
the central place to add other such instances.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>