Fiji Build: fix the infamous 'need multiple builds' bug
As noticed by both Stephan Saalfeld and Albert Cardona (and possibly a
million other people, but then they did not grace me with a bug report),
Fiji Build used to need several iterations until it would finally not
build anything any longer.

The reason: 1d37035(Fiji Build: mark targets as up-to-date if necessary).
This commit did not fully implement what it tried to do.

Originally, we wanted to commit all the generated .jar files, so it was
necessary to have a better up-to-date check than verifying that two
.jar files are byte-wise identical.  Remember, .jar files are just
glorified .zip files, and therefore contain timestamps.  When two
different developers compile the same .jar file, the result is therefore
most likely _not_ byte-wise identical.

Therefore, Fiji Build checks the contents of .jar files for identity
(because the .class files _are_ byte-wise identical).

However, the check was a bit costly, so this developer had the idea to
update the timestamps after such a check succeeded.

But we cannot update the timestamps right away, otherwise rules which
have the just-updated file as prerequisite would believe that they
need to be built even if the only change was the mtime of the file.

So we need to mark all the targets which think that they are out-of-date
just after a build succeeded (because then the targets are by definition
up-to-date).

This change actually _requires_ that the up-to-date check does not fail
when the mtimes are identical, because that will now happen pretty
regularly.

Of course, this whole strategy does not play well together with partial
builds that have common prerequisites (e.g. both Jython_Interpreter.jar
and JRuby_Scripting.jar depending on fiji-scripting.jar).  But hey,
life's not always a bowl of cherries.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>