update-fiji: respect dependencies
When uploading .jar files, care needs to be taken that dependencies are
uploaded as well, if they are locally modified.

To make things a bit less magic, the uploader will show the .jar files
that would be uploaded, too, and exit, unless the option '--auto' was
passed to the script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>