Fiji Build: when expanding globs, sort the result
The Fiji Updater calculates the checksums of the contents of .jar files
so that different timestamps will not make the checksums different.

But for the same reason, we must have consistent ordering of the file
names in the .jar files.  We should not really order the complete list,
as the order of some files might be crucial.  But we can order the
expanded glob lists, and the rest of the order is determined by
the Fakefile, so the overall order will be deterministic.

Noticed while writing documentation about the internals of the Fiji
Updater (so it seems documentation is also good for finding bugs).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>