open-in-gitweb.sh: use xdg-open to open the web pages
Not everybody has (or wants) Firefox.  The standard way on Linux to use the
user's preferred browser is to use xdg-open, which does all the
heavy-lifting.

For MacOSX and Windows, I think the appropriate tool are called "open" and
"start", respectively, so let's fall back to those if 'xdg-open' cannot
be executed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>