Rewrite release-fiji.sh to make use of nightly-build.sh
While at it, do not pull from master, but rather push from the current
repository.  Also, build the .7z archives, but leave out the Live CD and
Live USB Image for now: they need more work (Stephan Preibisch requested
that the Fiji directory be outside of the compressed image so it can be
easily made an all-platforms Fiji).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>