Fix borked precompiled launchers
The recent change that let the environment override the setting of
JAVA_HOME borked the precompiled launchers: they all would start
looking for the Linux-64 bit version of Java, and if that was
present, fail.

Besides, the recent upgrade to a newer Win64 cross compiler now
requires us to add the -static flag.  Otherwise we would have
to ship libgcc_s_sjlj-1.dll...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>