Fiji Updater: do not use the Prefs to store the mtime of db.xml.gz
There was a problem with the way we handled things: Prefs is user-wide,
while the mtime of db.xml.gz is installation wide.  So we step into
problems when the installation is either open to updates from all
users or when one user has several Fiji directories.

So just adjust the mtime of db.xml.gz instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>