Fiji launcher: ignore invalid JAVA_HOME
When JAVA_HOME is set, we do not force Fiji's JRE onto the user, but
assume that the user knows what she is doing.  However, that fails
when JAVA_HOME points somewhere over the rainbow.

So check that the directory JAVA_HOME refers to actually exists, and
ignore the JAVA_HOME setting otherwise.

Noticed by Albert Cardona.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>