release-fiji: only tag and upload to updater after user verified it
Instead of tagging and pushing before the files were tested, upload the
fiji-* files to /var/www/downloads/, and wait for the user to check that
everything works fine.

Only then proceed to tag and push (but only the tag, because 'master'
could have advanced in the meantime).

While at it, build the live images, too, and after tagging, also upload
the relevant files to the updater.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>