Fiji launcher(MacOSX): avoid overwriting argv[1] for non-Tiger systems
For hysterical raisins, we have to launch 32-bit all the time, only to
fall back to 64-bit when we're on a 64-bit system younger than Tiger.

The way we do that is to replace the "-tiger" suffix by a "-macosx"
suffix.  The astute reader will see that the replacement is one
byte longer than the suffix it replaces, and indeed, this led to an
overwriting of the next byte, which happens to be the first byte of
argv[1].  In effect, we replaced the first argument by an empty string.

To avoid that, let's just allocate a big-enough buffer and use that
as the new argv[0].

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>