Rework what is shown when opening the .dmg
For macros and plugins, we have options in Fiji to install them more
conveniently than to drag 'n drop them to the plugins folder
(Plugins>Install Plugin...), so do not offer the symbolic links anymore.

However, we generate a background image that tells the user to move the
Fiji application into the Applications folder, as suggested by Nico
Stuurman and Daniel James White.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>