Stack_Manipulation: new plugin: Slice Keeper
This plugin allows one to keep every n-th slice of a stack.

A request from Guenter Giese, with OKs from Tony Collins. The plugin
belonged to the MacBiophotonics package so far.
Tony told us it modified it from Wayne Rasband's Slice remover plugin,
and that credit for this one should go to him.

Signed-off-by: Jean-Yves Tinevez <jean-yves.tinevez@pasteur.fr>