Teach Fiji Build to use a separate output directory
When the variable "buildDir" is set in the Fakefile, it will be used as
target for the compiled .class files (similar to how things are done
with Ant).

To keep the .class files of different targets separate, they are stored
in subdirectories named after the respective target.

Requested by Stephan Saalfeld.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>