Refresh Javas: fall back to launching main() method
It is natural to assume that the user meant to call the static main
method of a class if it does not implement the PlugIn interface nor
the PlugInFilter interface.

Now that we do the class loading and executing ourselves, it is easy
to do that.

Wished for (a long time ago) by Albert Cardona.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>